Register/Login/Member's Area
Created the register.php where user goes to register their credentials.
 register.php runs validation.php to check if a user’s email already
exist in the database.

The login.php page is where user logs in.  It runs the checkuser.php
page.  If a user exist, and the credential is wrong, it will tell the
user that their credentials are wrong.  If they are correct, they are
taken to the home.php page.

home.php page is a protected area where only users who have
successfully logged in may access.  Page can not be access directly
without logging in.

logout.php is a page that a user can click on to destroy the existing
session and logging the user out of their credentials.

connection.php has database credentials.